requirements.txt
Flask==3.0.1
gunicorn==20.1.0
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
python-dotenv==0.19.0